Fix period cluttering with content & remove emphatic style (#4)
* Add margin between span of time and content

* Remove bold style from education & experience

The have nothing special to be emphasized.

* Use full html instead of mixing with markdown

Though they have the same effect, it's messy to preview a file
that has some sections mixing HTML with markdown since some of the
special characters can't be interpreted properly, such as bold (**)
and line break (\).

* Rearrange elements for better markdown previewing

So that location and time are display after institution and job
title respectively. No effect on actual rendering.

* Remove italic style from education & experience

* Update Chinese about page

* Make location stick to inst. & period stick to title

This is a reversion. The change in bd0eb6e982a5e05c3a24acd9d4a5093aa2034a84 do affect the rendering.

* Rename css class